Checkout transaction.  Define all transaction sources and destinations as interfaces so we won't need to make a breaking change if/when we support more sources.  Use Doubles instead of Floats.

+++ src/main/java/com/lightrail/model/Value.java
@@ -21,7 +21,7 @@ public class Value {
     public Boolean canceled;
     public RedemptionRule redemptionRule;
     public Boolean discount;
-    public Float discountSellerLiability;
+    public Double discountSellerLiability;
     public Date startDate;
     public Date endDate;
     public Map<String, Object> metadata;
